+++ test/mitmproxy/proxy/test_mode_servers.py
@@ -2,6 +2,8 @@ import asyncio
 from typing import cast
 from unittest.mock import AsyncMock, MagicMock, Mock
 
+import pytest
+
 from mitmproxy.net import udp
 from mitmproxy.proxy.mode_servers import DnsInstance, ServerInstance, DtlsInstance
 from mitmproxy.test import taddons
@@ -39,6 +41,24 @@ async def test_tcp_start_stop():
         assert await tctx.master.await_log("Stopped regular proxy server.")
 
 
+async def test_tcp_start_error():
+    manager = MagicMock()
+
+    with taddons.context() as tctx:
+        inst = ServerInstance.make("regular@127.0.0.1:0", manager)
+        await inst.start()
+        assert await tctx.master.await_log("proxy listening")
+        port = inst.listen_addrs[0][1]
+        inst2 = ServerInstance.make(f"regular@127.0.0.1:{port}", manager)
+        with pytest.raises(OSError, match=f"proxy failed to listen on 127\\.0\\.0\\.1:{port}"):
+            await inst2.start()
+        tctx.options.listen_host = "127.0.0.1"
+        tctx.options.listen_port = port
+        inst3 = ServerInstance.make(f"regular", manager)
+        with pytest.raises(OSError):
+            await inst3.start()
+
+
 async def test_udp_start_stop():
     manager = MagicMock()
 
@@ -59,6 +79,19 @@ async def test_udp_start_stop():
 
         await inst.stop()
         assert await tctx.master.await_log("Stopped")
+
+
+async def test_udp_start_error():
+    manager = MagicMock()
+
+    with taddons.context() as tctx:
+        inst = ServerInstance.make("dns@127.0.0.1:0", manager)
+        await inst.start()
+        assert await tctx.master.await_log("server listening")
+        port = inst.listen_addrs[0][1]
+        inst2 = ServerInstance.make(f"dns@127.0.0.1:{port}", manager)
+        with pytest.raises(OSError, match=f"server failed to listen on 127\\.0\\.0\\.1:{port}"):
+            await inst2.start()
 
 
 async def test_dtls_start_stop(monkeypatch):
@@ -82,12 +115,12 @@ async def test_udp_connection_reuse(monkeypatch):
     manager.connections = {}
 
     monkeypatch.setattr(udp, "DatagramWriter", MagicMock())
-    monkeypatch.setattr(DnsInstance, "handle_dns_connection", AsyncMock())
+    monkeypatch.setattr(DnsInstance, "handle_udp_connection", AsyncMock())
 
     with taddons.context():
         inst = cast(DnsInstance, ServerInstance.make("dns", manager))
-        inst.handle_dns_datagram(MagicMock(), b"\x00\x00\x01", ("remoteaddr", 0), ("localaddr", 0))
-        inst.handle_dns_datagram(MagicMock(), b"\x00\x00\x02", ("remoteaddr", 0), ("localaddr", 0))
+        inst.handle_udp_datagram(MagicMock(), b"\x00\x00\x01", ("remoteaddr", 0), ("localaddr", 0))
+        inst.handle_udp_datagram(MagicMock(), b"\x00\x00\x02", ("remoteaddr", 0), ("localaddr", 0))
         await asyncio.sleep(0)
 
         assert len(inst.manager.connections) == 1
@@ -98,12 +131,12 @@ async def test_dtls_connection_reuse(monkeypatch):
     manager.connections = {}
 
     monkeypatch.setattr(udp, "DatagramWriter", MagicMock())
-    monkeypatch.setattr(DtlsInstance, "handle_dtls_connection", AsyncMock())
+    monkeypatch.setattr(DtlsInstance, "handle_udp_connection", AsyncMock())
 
     with taddons.context():
         inst = cast(DtlsInstance, ServerInstance.make("dtls:reverse:127.0.0.1:0", manager))
-        inst.handle_dtls_datagram(MagicMock(), b"\x00\x00\x01", ("remoteaddr", 0), ("localaddr", 0))
-        inst.handle_dtls_datagram(MagicMock(), b"\x00\x00\x02", ("remoteaddr", 0), ("localaddr", 0))
+        inst.handle_udp_datagram(MagicMock(), b"\x00\x00\x01", ("remoteaddr", 0), ("localaddr", 0))
+        inst.handle_udp_datagram(MagicMock(), b"\x00\x00\x02", ("remoteaddr", 0), ("localaddr", 0))
         await asyncio.sleep(0)
 
         assert len(inst.manager.connections) == 1