+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased: mitmproxy next
 
+* Include server information in bind/listen errors.
+  ([#5495](https://github.com/mitmproxy/mitmproxy/pull/5495), @meitinger)
 * Include information about lazy connection_strategy in related errors.
   ([#5465](https://github.com/mitmproxy/mitmproxy/pull/5465), @meitinger, @mhils)
 * Add UDP layer and flow support.